# src/api-serverless/openapi.yaml
openapi: 3.0.3
info:
  title: Seize API
  description: This is the API interface description for the most commonly used operations in Seize API.  Some modifying endpoints require an authentication token.   We are in the process of documenting all Seize APIs.   If there is an API that you need, please ping us in Discord and we will aim to prioritize its documentation.
  version: 1.0.0
servers:
  - url: /api
paths:
  /auth/nonce:
    get:
      tags:
        - Authentication
      summary: Get a message to sign with your web3 wallet
      operationId: getNonce
      parameters:
        - name: signer_address
          in: query
          description: Your wallet address
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NonceResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate and get JWT token
      operationId: getAuthToken
      parameters:
        - name: signer_address
          in: query
          description: Your wallet address
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: 'Invalid request'
...
  /drops:
    get:
      tags:
        - Drops
      summary: Get latest drops
      operationId: getLatestDrops
      parameters:
        - name: limit
          in: query
          description: How many drops to return (10 by default)
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 20
        - name: curation_criteria_id
          in: query
          description: Drops by authors that match the supplied criteria
          required: false
          schema:
            type: string
        - name: wave_id
          in: query
          description: Drops in wave with given ID
          required: false
          schema:
            type: string
        - name: serial_no_less_than
          in: query
          description: Used to find older drops
          required: false
          schema:
            type: integer
            format: int64
        - name: min_part_id
          in: query
          description: Minimum ID of the parts wrapped in drops
          required: false
          schema:
            type: number
            format: int64
            minimum: 1
        - name: max_part_id
          in: query
          description: Maximum ID of the parts wrapped in drops
          required: false
          schema:
            type: number
            format: int64
            minimum: 1
        - name: context_profile
          in: query
          description: If supplied, each drop has some context related to given profile. It can be wallet address, ENS, handle or profile ID.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Drop"
    post:
      tags:
        - Drops
      summary: Create a drop
      description: "Requires the user to be authenticated"
      operationId: createDrop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDropRequest"
      responses:
        "201":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drop"
        "400":
          description: "Invalid request"
  /drops/{dropId}:
    get:
      tags:
        - Drops
      summary: Get drop by ID
      operationId: getDropById
      parameters:
        - name: dropId
          in: path
          required: true
          schema:
            type: string
        - name: context_profile
          in: query
          description: If supplied drop has some context related to given profile. It can be wallet address, ENS, handle or profile ID.
          required: false
          schema:
            type: string
        - name: min_part_id
          in: query
          description: Minimum ID of the parts wrapped in drops
          required: false
          schema:
            type: number
            format: int64
            minimum: 1
        - name: max_part_id
          in: query
          description: Maximum ID of the parts wrapped in drops
          required: false
          schema:
            type: number
            format: int64
            minimum: 1
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drop"
        "404":
          description: Drop not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Drop:
      type: object
      required:
        - id
        - serial_no
        - wave_id
        - author
        - created_at
        - referenced_nfts
        - mentioned_users
        - metadata
        - parts
        - parts_count
        - vote
        - top_voters
        - voters_count
        - voting_logs_count
        - title
        - context_profile_context
      properties:
        id:
          type: string
        serial_no:
          description: Sequence number of the drop in Seize
          type: number
          format: int64
        wave_id:
          type: string
        author:
          $ref: "#/components/schemas/ProfileMin"
        created_at:
          description: Time when the drop was created in milliseconds since 1-1-1970 00:00:00.0 UTC
          type: number
          format: int64
        title:
          type: string
          maxLength: 250
          nullable: true
        parts:
          type: array
          items:
            $ref: "#/components/schemas/DropPart"
        parts_count:
          description: "Number of drops in the storm"
          type: number
          format: int64
        referenced_nfts:
          type: array
          items:
            $ref: "#/components/schemas/DropReferencedNFT"
        mentioned_users:
          type: array
          items:
            $ref: "#/components/schemas/DropMentionedUser"
        metadata:
          type: array
          items:
            $ref: "#/components/schemas/DropMetadata"
        vote:
          type: number
          format: int64
        top_voters:
          type: array
          items:
            $ref: "#/components/schemas/DropVoter"
        voters_count:
          type: number
          format: int64
        voting_logs_count:
          type: number
          format: int64
        context_profile_context:
          anyOf:
            - type: "null"
            - $ref: "#/components/schemas/DropContextProfileContext"
    QuotedDrop:
      type: object
      required:
        - drop_id
        - drop_part_id
      properties:
        drop_id:
          type: string
        drop_part_id:
          type: number
          format: int64
    DropPart:
      type: object
      required:
        - part_id
        - media
        - discussion_comments_count
        - quotes_count
        - quoted_drop
        - content
      properties:
        part_id:
          description: "Which drop in the storm is this drop"
          type: number
          format: int64
        content:
          type: string
          maxLength: 25000
          nullable: true
        media:
          type: array
          items:
            $ref: "#/components/schemas/DropMedia"
        quoted_drop:
          anyOf:
            - $ref: "#/components/schemas/QuotedDrop"
            - type: "null"
        discussion_comments_count:
          type: number
          format: int64
        quotes_count:
          type: number
          format: int64
        context_profile_context:
          anyOf:
            - type: "null"
            - $ref: "#/components/schemas/DropPartContextProfileContext"
    ProfileMin:
      type: object
      required:
        - id
        - handle
        - pfp
        - cic
        - rep
        - tdh
        - level
        - archived
      properties:
        id:
          type: string
          format: uuid
        handle:
          type: string
        pfp:
          type: string
          nullable: true
        cic:
          type: number
          format: int64
        rep:
          type: number
          format: int64
        tdh:
          type: number
          format: int64
        level:
          type: number
          format: int64
        archived:
          type: boolean
    DropReferencedNFT:
      type: object
      required:
        - contract
        - token
        - name
      properties:
        contract:
          type: string
          maxLength: 100
        token:
          type: string
          maxLength: 250
        name:
          type: string
          maxLength: 500
    DropMentionedUser:
      type: object
      required:
        - mentioned_profile_id
        - handle_in_content
      properties:
        mentioned_profile_id:
          type: string
          format: uuid
        handle_in_content:
          type: string
          maxLength: 100
        current_handle:
          type: string
          maxLength: 100
          nullable: true
    DropMetadata:
      type: object
      required:
        - data_key
        - data_value
      properties:
        data_key:
          type: string
          maxLength: 500
        data_value:
          type: string
          maxLength: 500
    DropVoter:
      type: object
      required:
        - profile
        - vote
      properties:
        profile:
          $ref: "#/components/schemas/ProfileMin"
        vote:
          type: number
          format: int64
...
    DropContextProfileContext:
      type: object
      required:
        - vote
        - total_credit
      properties:
        vote:
          type: number
          format: int64
        total_credit:
          type: number
          format: int64
    DropPartContextProfileContext:
      type: object
      required:
        - discussion_comments_count
        - quotes_count
      properties:
        discussion_comments_count:
          type: number
          format: int64
        quotes_count:
          type: number
          format: int64
    CreateDropPart:
      type: object
      required:
        - media
      properties:
        content:
          type: string
          maxLength: 25000
          nullable: true
        quoted_drop:
          anyOf:
            - $ref: "#/components/schemas/QuotedDrop"
            - type: "null"
        media:
          type: array
          items:
            $ref: "#/components/schemas/DropMedia"
    CreateDropRequest:
      type: object
      required:
        - wave_id
        - referenced_nfts
        - mentioned_users
        - metadata
        - parts
      properties:
        wave_id:
          type: string
        title:
          type: string
          maxLength: 250
          nullable: true
        parts:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateDropPart"
        referenced_nfts:
          type: array
          items:
            $ref: "#/components/schemas/DropReferencedNFT"
        mentioned_users:
          type: array
          items:
            $ref: "#/components/schemas/DropMentionedUser"
        metadata:
          type: array
          items:
            $ref: "#/components/schemas/DropMetadata"
    DropMedia:
      type: object
      required:
        - url
        - mime_type
      properties:
        url:
          type: string
          maxLength: 2000
        mime_type:
          type: string
          maxLength: 100
...
    NonceResponse:
      type: object
      required:
        - nonce
        - server_signature
      properties:
        nonce:
          type: string
        server_signature:
          type: string
    LoginRequest:
      type: object
      required:
        - client_signature
        - server_signature
      properties:
        client_signature:
          type: string
        server_signature:
          type: string
        role:
          type: string
    LoginResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
...
security:
  - bearerAuth: []