# src/api-serverless/src/generated/models/Drop.ts
/**
 * Seize API
 * This is the API interface description for the most commonly used operations in Seize API.  Some modifying endpoints require an authentication token.   We are in the process of documenting all Seize APIs.   If there is an API that you need, please ping us in Discord and we will aim to prioritize its documentation.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DropContextProfileContext } from '../models/DropContextProfileContext';
import { DropMentionedUser } from '../models/DropMentionedUser';
import { DropMetadata } from '../models/DropMetadata';
import { DropPart } from '../models/DropPart';
import { DropReferencedNFT } from '../models/DropReferencedNFT';
import { DropVoter } from '../models/DropVoter';
import { ProfileMin } from '../models/ProfileMin';
import { HttpFile } from '../http/http';

export class Drop {
    'id': string;
    /**
    * Sequence number of the drop in Seize
    */
    'serial_no': number;
    'wave_id': string;
    'author': ProfileMin;
    /**
    * Time when the drop was created in milliseconds since 1-1-1970 00:00:00.0 UTC
    */
    'created_at': number;
    'title': string | null;
    'parts': Array<DropPart>;
    /**
    * Number of drops in the storm
    */
    'parts_count': number;
    'referenced_nfts': Array<DropReferencedNFT>;
    'mentioned_users': Array<DropMentionedUser>;
    'metadata': Array<DropMetadata>;
    'vote': number;
    'top_voters': Array<DropVoter>;
    'voters_count': number;
    'voting_logs_count': number;
    'context_profile_context': DropContextProfileContext | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string",
            "format": ""
        },
        {
            "name": "serial_no",
            "baseName": "serial_no",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "wave_id",
            "baseName": "wave_id",
            "type": "string",
            "format": ""
        },
        {
            "name": "author",
            "baseName": "author",
            "type": "ProfileMin",
            "format": ""
        },
        {
            "name": "created_at",
            "baseName": "created_at",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string",
            "format": ""
        },
        {
            "name": "parts",
            "baseName": "parts",
            "type": "Array<DropPart>",
            "format": ""
        },
        {
            "name": "parts_count",
            "baseName": "parts_count",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "referenced_nfts",
            "baseName": "referenced_nfts",
            "type": "Array<DropReferencedNFT>",
            "format": ""
        },
        {
            "name": "mentioned_users",
            "baseName": "mentioned_users",
            "type": "Array<DropMentionedUser>",
            "format": ""
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "Array<DropMetadata>",
            "format": ""
        },
        {
            "name": "vote",
            "baseName": "vote",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "top_voters",
            "baseName": "top_voters",
            "type": "Array<DropVoter>",
            "format": ""
        },
        {
            "name": "voters_count",
            "baseName": "voters_count",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "voting_logs_count",
            "baseName": "voting_logs_count",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "context_profile_context",
            "baseName": "context_profile_context",
            "type": "DropContextProfileContext",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return Drop.attributeTypeMap;
    }

    public constructor() {
    }
}